pyproject.toml
[tool.poetry]
name = "cgn-ec-syslog-proxy"
version = "0.1.0"
description = ""
authors = ["BSpendlove <8687668+BSpendlove@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.6"
pydantic-settings = "^2.9.1"
pydantic = "^2.11.4"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
